+++ fabric/src/testmod/java/net/frozenblock/lib/testmod/registry/FTLootTables.java
@@ -17,6 +17,7 @@
 
 package net.frozenblock.lib.testmod.registry;
 
+import lombok.experimental.UtilityClass;
 import net.frozenblock.lib.item.api.loot.LootTableModification;
 import net.frozenblock.lib.item.impl.loot.MutableLootTable;
 import net.frozenblock.lib.testmod.FrozenTestMain;
@@ -25,26 +26,24 @@ import net.minecraft.world.level.storage.loot.BuiltInLootTables;
 import net.minecraft.world.level.storage.loot.functions.SetItemCountFunction;
 import net.minecraft.world.level.storage.loot.providers.number.UniformGenerator;
 
+@UtilityClass
 public final class FTLootTables {
-	private FTLootTables() {
-		throw new UnsupportedOperationException("FTLootTables only supports static declarations.");
-	}
 
 	public static void init() {
 		FrozenTestMain.LOGGER.info("Registering Loot Table Modifications for FrozenTest.");
 		//BONUS CHEST
 		LootTableModification.editTable(
 			BuiltInLootTables.SPAWN_BONUS_CHEST, false,
-			(id, mutableLootTable) -> mutableLootTable.modifyPools(
+			(id, mutableLootTable, registries) -> mutableLootTable.modifyPools(
 				MutableLootTable.has(Items.ACACIA_LOG),
-				(lootPool) -> lootPool.add(Items.DIAMOND_BLOCK, 3, SetItemCountFunction.setCount(UniformGenerator.between(1.0F, 3.0F)))
+				(lootPool) -> lootPool.add(Items.DIAMOND_BLOCK, 3, SetItemCountFunction.setCount(UniformGenerator.between(1F, 3F)))
 			)
 		);
 
 		//Cold Ocean Ruin Archaeology
 		LootTableModification.editTable(
 			BuiltInLootTables.OCEAN_RUIN_COLD_ARCHAEOLOGY, false,
-			(id, mutableLootTable) -> mutableLootTable.modifyPools(
+			(id, mutableLootTable, registries) -> mutableLootTable.modifyPools(
 				MutableLootTable.has(Items.MOURNER_POTTERY_SHERD),
 				(lootPool) -> lootPool.replace(Items.MOURNER_POTTERY_SHERD, Items.FLOW_POTTERY_SHERD)
 			)